backend/requirements.txt
asgiref
certifi
cffi
charset-normalizer
cryptography
defusedxml
django
django-cors-headers
djangorestframework
djangorestframework-jwt
djangorestframework-simplejwt
djoser
idna
oauthlib
pycparser
pyjwt
python3-openid
requests
requests-oauthlib
social-auth-app-django
social-auth-core
sqlparse
urllib3
